requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
drf-spectacular